requirements.txt
Django==5.2.12

djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
pytest-django==4.12.0

django-filter==25.2
drf-spectacular==0.29.0

psycopg2-binary==2.9.9
python-dotenv==1.0.1

gunicorn==22.0.0
pillow==11.1.0
